requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests==2.31.0
typer==0.9.0
rich==13.7.0
pytest==8.0.0
httpx==0.26.0
python-multipart==0.0.9
psutil==5.9.8
